+++ lib/online_migrations/command_recorder.rb
@@ -153,4 +153,157 @@ module OnlineMigrations
         [:add_text_limit_constraint, args]
       end
   end
+
+  # @private
+  # New format was introduced in https://github.com/rails/rails/pull/58239.
+  module CommandRecorder82
+    REVERSIBLE_AND_IRREVERSIBLE_METHODS = [
+      :update_column_in_batches,
+      :initialize_column_rename,
+      :initialize_columns_rename,
+      :revert_initialize_column_rename,
+      :revert_initialize_columns_rename,
+      :finalize_column_rename,
+      :finalize_columns_rename,
+      :revert_finalize_column_rename,
+      :revert_finalize_columns_rename,
+      :initialize_table_rename,
+      :revert_initialize_table_rename,
+      :finalize_table_rename,
+      :revert_finalize_table_rename,
+      :swap_column_names,
+      :add_column_with_default,
+      :add_not_null_constraint,
+      :remove_not_null_constraint,
+      :add_text_limit_constraint,
+      :remove_text_limit_constraint,
+      :add_reference_concurrently,
+      :change_column_type_in_background,
+      :enqueue_background_data_migration,
+      :remove_background_data_migration,
+
+      # column type change helpers
+      :initialize_column_type_change,
+      :initialize_columns_type_change,
+      :revert_initialize_column_type_change,
+      :revert_initialize_columns_type_change,
+      :backfill_column_for_type_change,
+      :backfill_columns_for_type_change,
+      :finalize_column_type_change,
+      :finalize_columns_type_change,
+      :revert_finalize_column_type_change,
+      :cleanup_column_type_change,
+      :cleanup_columns_type_change,
+    ]
+
+    REVERSIBLE_AND_IRREVERSIBLE_METHODS.each do |method|
+      class_eval <<-RUBY, __FILE__, __LINE__ + 1
+        def #{method}(*args, **kwargs, &block)          # def create_table(*args, **kwargs, &block)
+          record(:"#{method}", args, kwargs, &block)    #   record(:create_table, args, kwargs, &block)
+        end                                             # end
+      RUBY
+    end
+
+    private
+      module StraightReversions
+        {
+          initialize_column_rename:           :revert_initialize_column_rename,
+          initialize_columns_rename:          :revert_initialize_columns_rename,
+          finalize_column_rename:             :revert_finalize_column_rename,
+          finalize_columns_rename:            :revert_finalize_columns_rename,
+          initialize_table_rename:            :revert_initialize_table_rename,
+          finalize_table_rename:              :revert_finalize_table_rename,
+          add_not_null_constraint:            :remove_not_null_constraint,
+          initialize_column_type_change:      :revert_initialize_column_type_change,
+          initialize_columns_type_change:     :revert_initialize_columns_type_change,
+          finalize_column_type_change:        :revert_finalize_column_type_change,
+          finalize_columns_type_change:       :revert_finalize_columns_type_change,
+        }.each do |cmd, inv|
+          [[inv, cmd], [cmd, inv]].each do |method, inverse|
+            class_eval <<-RUBY, __FILE__, __LINE__ + 1
+              def invert_#{method}(args, kwargs, &block)    # def invert_create_table(args, kwargs, &block)
+                [:#{inverse}, args, kwargs, block]          #   [:drop_table, args, kwargs, block]
+              end                                           # end
+            RUBY
+          end
+        end
+      end
+
+      include StraightReversions
+
+      def invert_add_reference_concurrently(args, kwargs)
+        [:remove_reference, args, kwargs]
+      end
+
+      def invert_swap_column_names(args, kwargs)
+        table_name, column1, column2 = args
+        [:swap_column_names, [table_name, column2, column1], kwargs]
+      end
+
+      def invert_add_column_with_default(args, kwargs)
+        table_name, column_name, = args
+        [:remove_column, [table_name, column_name], kwargs]
+      end
+
+      def invert_revert_initialize_column_rename(args, kwargs)
+        _table, column, new_column = args
+        if !column || !new_column
+          raise ActiveRecord::IrreversibleMigration,
+            "revert_initialize_column_rename is only reversible if given a column and new_column."
+        end
+        [:initialize_column_rename, args, kwargs]
+      end
+
+      def invert_revert_initialize_columns_rename(args, kwargs)
+        _table, old_new_column_hash = args
+        if !old_new_column_hash
+          raise ActiveRecord::IrreversibleMigration,
+            "revert_initialize_columns_rename is only reversible if given a hash of old and new columns."
+        end
+        [:initialize_columns_rename, args, kwargs]
+      end
+
+      def invert_finalize_table_rename(args, kwargs)
+        _table_name, new_name = args
+        if !new_name
+          raise ActiveRecord::IrreversibleMigration,
+            "finalize_table_rename is only reversible if given a new_name."
+        end
+        [:revert_finalize_table_rename, args, kwargs]
+      end
+
+      def invert_revert_initialize_column_type_change(args, kwargs)
+        if !args[2]
+          raise ActiveRecord::IrreversibleMigration,
+            "revert_initialize_column_type_change is only reversible if given a new_type."
+        end
+        super
+      end
+
+      def invert_revert_initialize_columns_type_change(args, kwargs)
+        if args[1].empty?
+          raise ActiveRecord::IrreversibleMigration,
+            "revert_initialize_columns_type_change is only reversible if given a columns_and_types."
+        end
+        super
+      end
+
+      def invert_add_not_null_constraint(args, kwargs)
+        kwargs.delete(:validate)
+        [:remove_not_null_constraint, args, kwargs]
+      end
+
+      def invert_add_text_limit_constraint(args, kwargs)
+        kwargs.delete(:validate)
+        [:remove_text_limit_constraint, args, kwargs]
+      end
+
+      def invert_remove_text_limit_constraint(args, kwargs)
+        if !args[2]
+          raise ActiveRecord::IrreversibleMigration, "remove_text_limit_constraint is only reversible if given a limit."
+        end
+
+        [:add_text_limit_constraint, args, kwargs]
+      end
+  end
 end