+++ lib/online_migrations/schema_dumper.rb
@@ -15,7 +15,13 @@ module OnlineMigrations
 
   class WrappedConnection < SimpleDelegator
     def columns(table_name)
-      super.sort_by(&:name)
+      columns = super
+
+      if Utils.ar_version >= 8.2 && columns.is_a?(Hash)
+        columns.transform_values { |v| v.sort_by(&:name) }
+      else
+        columns.sort_by(&:name)
+      end
     end
   end
 end